Alien Bullet Collision
Added alien bullet collision for hitting bunkers and hitting the tank.
Added basic tank death. One issue was as a bullet was firing it's shape
changed, so if we do the bullets purely based on their shape, they can
appear to magically move through blocks, so for the algorithm is based
on overlapping with bunker blocks.