Enabled sound interrupt, now logic issues
I was successfully able to enable the interrupt to allow the FIFO to
signal when it was half empty. However, there seems to be an error in my
logic for activating the sounds. My thinking was to turn each sound to
active when it should happen in the game and then use the fillFIFO
function to fill the FIFO with the correct sounds every time the
interrupt happens. Once again, the debug is not working.